requirements.txt
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
plotly==5.15.0
dash==2.11.1
mysql-connector-python==8.1.0
numpy==1.24.3
